matrice.py -> linealge.py, et contient une grande quantité de fonctions lié à l'algèbre linéaire, ajout dans la doc
